feat(elements): Move loading into common
