+++ packages/elements/src/react/sign-up/index.ts
@@ -9,8 +9,6 @@ export { SignUpAction as Action } from './action';
 export { SignUpStrategy as Strategy } from './verifications';
 export { SignUpProvider as Provider, SignUpProviderIcon as ProviderIcon } from './providers';
 
-export { Loading } from './loading';
-
 /** @internal Internal use only */
 export const useSignUpActorRef_internal = SignUpRouterCtx.useActorRef;
